+++ CHANGELOG.md
@@ -2,6 +2,8 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+## v23.4.29
+
 ### Bugfixes
 
 * Fix decision making for simpler/faster programs